feat: add file not found handle

+++ shcdecryptor.sh
@@ -2,11 +2,17 @@
 
 if [[ $1 == "--installation-testing" ]]; then
   echo "ok"
+  exit
 fi
 
 for args in $@; do
+  if ! [[ -f $args ]]; then
+    echo "$args not found."
+    exit
+  fi
+
   $args > /dev/null & child=$!
-  sleep 0.0.3
+  sleep 0.03
   kill -STOP $child
   cat /proc/$child/cmdline | sed 's/.*\(#!\)/\1/; $d' > $args.decrypted.sh
   kill -TERM $child
